+++ roslibrust_common/Cargo.toml
@@ -7,6 +7,8 @@ edition = "2021"
 thiserror = "2.0"
 anyhow = "1.0"
 serde = { version = "1.0", features = ["derive"] }
+# Used for md5sum calculation
+md5 = "0.7"
 
 # THESE DEPENDENCIES WILL BE REMOVED
 # We're currently leaking these error types in the "generic error type"
